==> R/p2_WGCNA_03_11_16.Rmd <==
---
title: "Meta_WGCNA_GAP"
author: "Daniel Leirer"
date: "20/01/2016"
output: html_document
---


##Install and load Libraries:
```{r Libraries, message=FALSE,warning=FALSE}
rm(list = ls())
dev.off()
library(WGCNA);
library(flashClust)
library(tidyr)
library(dplyr)
```


##Set directories:
```{r directory}
topdir<-getwd()
data_dir <-"./P0_Characterise/output/"
P2_output_dir <-"./P2_WGCNA/output/"
P2_figs_dir <-"./P2_WGCNA/figs/"

```

##Load data and Settings
```{r data and settings}
<<<<<<< HEAD

=======
>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2

# The following setting is important, do not omit.
options(stringsAsFactors = FALSE);


#data="GAP_FEP_Full_Gene_Expression_Data.RData"
data="GAP_FEP_eset_linear_adj_Data.RData"

#load Gene expression data
load(paste(data_dir,data,sep=""))

GX_lumi_object<-eset_linear_adj

# remove duplicate probes and use gene symbols
gene<- t(exprs(GX_lumi_object))
probenames<-GX_lumi_object@featureData@data
colnames(gene) <- probenames$TargetID
gene_expression_matrix<-t(gene)
t_gene_expression_matrix<-gene



dim(t_gene_expression_matrix)

##Check for missing values
gsg = goodSamplesGenes(t_gene_expression_matrix, verbose = 3);
gsg$allOK
```

##Loading Phenotype Data
```{r load pheno}
names(pData(GX_lumi_object))
allTraits<-pData(GX_lumi_object)[,c(1,11:17,21:26,32:33,62,63)]

## Change Tech Names
colnames(allTraits)[13:16]<-c("RIN","SentrixBarcode","Nanodrop","AmplificationDate")
colnames(allTraits)

# Form a data frame analogous to expression data that will hold the clinical traits.
samples = rownames(t_gene_expression_matrix);
traitRows = match(samples, allTraits$sampleID);
datTraits = allTraits[traitRows, -1];
rownames(datTraits) = allTraits[traitRows, 1];
collectGarbage();
str(datTraits)
datTraits$Phenotype<-as.numeric(as.factor(datTraits$Phenotype))
datTraits$Gender<-as.numeric(as.factor(datTraits$Gender))
datTraits$Ethnicity<-as.numeric(as.factor(datTraits$Ethnicity))
datTraits$Tobacco<-as.numeric(as.factor(datTraits$Tobacco))
datTraits$Medication<-as.numeric(as.factor(datTraits$Medication))
datTraits$SentrixBarcode<-as.numeric(as.factor(datTraits$SentrixBarcode))
datTraits$AmplificationDate<-as.numeric(as.factor(datTraits$AmplificationDate))


##Add all medication comparisions
datTraits<-datTraits%>%mutate(
  AF_vs_Con=ifelse(Medication == 2,2,
            ifelse(Medication == 4,1,NA)),
  Ris_vs_Con=ifelse(Medication == 8,2,
            ifelse(Medication == 4,1,NA)),
  Ola_vs_Con=ifelse(Medication == 6,2,
            ifelse(Medication == 4,1,NA)),
  Ris_vs_AF=ifelse(Medication == 8,2,
            ifelse(Medication == 2,1,NA)),
  Ola_vs_AF=ifelse(Medication == 6,2,
            ifelse(Medication == 2,1,NA)),
  Ris_vs_Ola=ifelse(Medication == 6,2,
            ifelse(Medication == 8,1,NA)),
  Med_vs_AF_Con=ifelse(Medication == 2,1,
            ifelse(Medication == 4,1,2))
  )








```

## Plot completion of samples
```{r}

plotname<-"sample_dendrogram.jpeg"

jpeg(file=paste(P2_figs_dir,plotname,sep=""),width=1000,height=800,pointsize = 12)
# Re-cluster samples
sampleTree2 = flashClust(dist(t_gene_expression_matrix), method = "average")
# Convert traits to a color representation: white means low, red means high, grey means missing entry
traitColors = numbers2colors(datTraits, signed = FALSE);
# Plot the sample dendrogram and the colors underneath.
plotDendroAndColors(sampleTree2, traitColors,
groupLabels = names(datTraits),
main = "Sample dendrogram and trait heatmap")
dev.off()


```


```{r}
# Choose a set of soft-thresholding powers
powers = c(1:15)
# Call the network topology analysis function
sft = pickSoftThreshold(t_gene_expression_matrix, powerVector = powers, verbose = 5, networkType = "unsigned")


plotname2<-"scale_independence_plot.jpeg"
jpeg(file=paste(P2_figs_dir,plotname2,sep=""),width=1000,height=800,pointsize = 12)
# Plot the results:
par(mfrow = c(1,2));
cex1 = 0.9;
# Scale-free topology fit index as a function of the soft-thresholding power
plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
xlab="Soft Threshold (power)",ylab="Scale Free Topology Model Fit,signed R^2",type="n",
main = paste("Scale independence"));
text(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
labels=powers,cex=cex1,col="red");
# this line corresponds to using an R^2 cut-off of h
abline(h=0.90,col="red")
# Mean connectivity as a function of the soft-thresholding power
plot(sft$fitIndices[,1], sft$fitIndices[,5],
xlab="Soft Threshold (power)",ylab="Mean Connectivity", type="n",
main = paste("Mean connectivity"))
text(sft$fitIndices[,1], sft$fitIndices[,5], labels=powers, cex=cex1,col="red")
dev.off()
par(mfrow = c(1,1));
```

```{r}
## Automatic
##net = blockwiseModules(t_gene_expression_matrix, power = 6, TOMType = "unsigned", minModuleSize = 10, reassignThreshold = 0, mergeCutHeight = 0.25, numericLabels = TRUE, pamRespectsDendro= FALSE, saveTOMs = TRUE, saveTOMFileBase = "femaleMouseTOM", verbose = 3)

# 6.1 Set soft power beta=6
softPower = 6;
adjacency = adjacency(t_gene_expression_matrix, power = softPower);


# 6.2 Turn adjacency into topological overlap
TOM = TOMsimilarity(adjacency,TOMType="unsigned");
dissTOM = 1-TOM

# 6.3  Call the hierarchical clustering function
geneTree = hclust(as.dist(dissTOM), method = "average");


# Plot the resulting clustering tree (dendrogram)
clust_tree<-"Clustering_tree.png"
png(paste(P2_figs_dir,clust_tree,sep=""))
plot(geneTree, xlab="", sub="", main = "Gene clustering on TOM-based dissimilarity",
     labels = FALSE, hang = 0.04);
dev.off()


# 6.4 Module identification

# We like large modules, so we set the minimum module size relatively high:
minModuleSize = 20;
# Module identification using dynamic tree cut:
dynamicMods = cutreeDynamic(dendro = geneTree, distM = dissTOM, deepSplit = 2, pamRespectsDendro = FALSE, minClusterSize = minModuleSize);
table(dynamicMods)

# 6.5 Convert numeric lables into colors and plot the dendrograms and colors underneath
dynamicColors = labels2colors(dynamicMods)
table(dynamicColors)

clust_tree<-"Dendrorogram.png"
png(paste(P2_figs_dir,clust_tree,sep=""))
plotDendroAndColors(geneTree, dynamicColors, "Dynamic Tree Cut", dendroLabels = FALSE, hang = 0.03, addGuide = TRUE, guideHang = 0.05, main = "Gene dendrogram and module colors")
dev.off()


# 6.6  Calculate eigengenes
MEList = moduleEigengenes(t_gene_expression_matrix, colors = dynamicColors)
MEs = MEList$eigengenes
# Calculate dissimilarity of module eigengenes
MEDiss = 1-cor(MEs);
# Cluster module eigengenes
METree = hclust(as.dist(MEDiss), method = "average");

setwd(P2_figs_dir)
# 6.7  Plot the results and see whether some Modules are highly correlated and therefore would be better to be merged
# Plot the cut line into the dendrogram
png("ME clustering.png")
plot(METree, main = "Clustering of module eigengenes",xlab = "", sub = "")
MEDissThres = 0.2 # we set the threshold of 0.25 ( scale is 0 to 1 - the smallest the value the more correlated)
abline(h=MEDissThres, col = "red")
dev.off()


#6.8  Call an automatic merging function to merge Modules which are very similar (using the MEDissThres threshold )
merge = mergeCloseModules(t_gene_expression_matrix, dynamicColors, cutHeight = MEDissThres, verbose = 3)
# The merged module colors
mergedColors = merge$colors;
# Eigengenes of the new merged modules:
mergedMEs = merge$newMEs;



#6.9  Plot the dendrograms and colors underneath based on the merging of Modules
png("Supplementary_Plots_geneDendro-UPDATED.png",width = 1000, height=800)
plotDendroAndColors(geneTree, cbind(dynamicColors, mergedColors),c("Dynamic Tree Cut", "Merged dynamic"), dendroLabels = FALSE, hang = 0.03, addGuide = TRUE, guideHang = 0.05)
dev.off()

png("Supplementary_Plots_geneDendroMerged.png",width = 1000, height=800)
plotDendroAndColors(geneTree, cbind(mergedColors),c("Merged dynamic"), dendroLabels = FALSE, hang = 0.03, addGuide = TRUE, guideHang = 0.05)
dev.off()

pdf("Supplementary_Figure_1_Modules.pdf")
plotDendroAndColors(geneTree, cbind(mergedColors),c("Modules"), dendroLabels = FALSE, hang = 0.03, addGuide = TRUE, guideHang = 0.05)
dev.off()



#6.10 Rename the new mergedColors to moduleColors (based on the updated merging)
moduleColors = mergedColors
# Construct numerical labels corresponding to the colors
colorOrder = c("grey", standardColors(50));
moduleLabels = match(moduleColors, colorOrder)-1;
MEs = mergedMEs;


setwd(topdir)
setwd(P2_output_dir)
write.csv(table(moduleColors), file = "Table_merged_modules.csv")
setwd(topdir)

```





```{r}
# Define numbers of genes and samples

nGenes = ncol(t_gene_expression_matrix);
nSamples = nrow(t_gene_expression_matrix);

names(datTraits)

# Recalculate MEs with color labels
MEs0 = moduleEigengenes(t_gene_expression_matrix, moduleColors)$eigengenes
MEs = orderMEs(MEs0)
moduleTraitCor = cor(MEs, datTraits, use = "p");
moduleTraitPvalue = corPvalueStudent(moduleTraitCor, nSamples);


```

```{r}

<<<<<<< HEAD

=======
setwd(P2_figs_dir)
>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2
jpeg(file="Module_Trait_Relationships.jpeg",width=1000,height=900,pointsize = 20)
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = names(datTraits),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()

pdf("Supplementary_Figure_2_Module_Trait_Relationships.pdf")
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = names(datTraits),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()

names(datTraits)
datTraits[,c(1,5,6,8:11)]
moduleTraitCor = cor(MEs, datTraits[,c(1,5,8:11)], use = "p");
moduleTraitPvalue = corPvalueStudent(moduleTraitCor, nSamples);
<<<<<<< HEAD
=======


jpeg(file="Module_Trait_Relationships_reduced.jpeg",width=1000,height=900,pointsize = 20)
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = colnames(moduleTraitCor),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()

pdf("Supplementary_Figure_2_Module_Trait_Relationships_reduced.pdf")
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = colnames(moduleTraitCor),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()

names(datTraits)
datTraits[,c(1,18:24,8:11,5,6)]
moduleTraitCor = cor(MEs, datTraits[,c(1,18:24,8:11,5,6)], use = "p");
moduleTraitPvalue = corPvalueStudent(moduleTraitCor, nSamples);


jpeg(file="Module_Trait_Relationships_med.jpeg",width=1000,height=900,pointsize = 20)
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = colnames(moduleTraitCor),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()
>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2

pdf("Supplementary_Figure_2_Module_Trait_Relationships_med.pdf")
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = colnames(moduleTraitCor),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()

jpeg(file="Module_Trait_Relationships_reduced.jpeg",width=1000,height=900,pointsize = 20)
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = colnames(moduleTraitCor),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()

pdf("Supplementary_Figure_2_Module_Trait_Relationships_reduced.pdf")
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = colnames(moduleTraitCor),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()

names(datTraits)
datTraits[,c(1,18:24,8:11,5,6)]
moduleTraitCor = cor(MEs, datTraits[,c(1,18:24,8:11,5,6)], use = "p");
moduleTraitPvalue = corPvalueStudent(moduleTraitCor, nSamples);


jpeg(file="Module_Trait_Relationships_med.jpeg",width=1000,height=900,pointsize = 20)
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = colnames(moduleTraitCor),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()

pdf("Supplementary_Figure_2_Module_Trait_Relationships_med.pdf")
# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
xLabels = colnames(moduleTraitCor),
yLabels = names(MEs),
ySymbols = names(MEs),
colorLabels = FALSE,
colors = blueWhiteRed(50),
textMatrix = textMatrix,
setStdMargins = FALSE,
cex.text = 0.5,
zlim = c(-1,1),
main = paste("Module-trait relationships"))
dev.off()


getwd()
```



```{r}
# Define variable weight containing the weight column of datTrait
psychosis = as.data.frame(datTraits$Phenotype);
names(psychosis) = "psychosis"
# names (colors) of the modules
modNames = substring(names(MEs), 3)

geneModuleMembership = as.data.frame(cor(t_gene_expression_matrix, MEs, use = "p"));
MMPvalue = as.data.frame(corPvalueStudent(as.matrix(geneModuleMembership), nSamples));

names(geneModuleMembership) = paste("MM", modNames, sep="");
names(MMPvalue) = paste("p.MM", modNames, sep="");

geneTraitSignificance = as.data.frame(cor(t_gene_expression_matrix, psychosis, use = "p"));
GSPvalue = as.data.frame(corPvalueStudent(as.matrix(geneTraitSignificance), nSamples));

names(geneTraitSignificance) = paste("GS.", names(psychosis), sep="");
names(GSPvalue) = paste("p.GS.", names(psychosis), sep="");
```

```{r}

pdf("ModuleMembership.pdf",height=8,width=8)
for (c in 1:length(modNames)){
  module =modNames[c]
  moduleGenes2 = moduleColors==modNames[c]
  verboseScatterplot(abs(geneModuleMembership[moduleGenes2,c]),
                     abs(geneTraitSignificance[moduleGenes2, 1]),
xlab = paste("Module Membership in", module, "module"),
ylab = "Gene significance for psychosis",
main = paste("Module membership vs. gene significance\n"),
cex.main = 1.2, cex.lab = 1.2, cex.axis = 1.2, col = module)
}; 
dev.off()

pdf("ModuleMembership_above_median.pdf",height=8,width=8)
for (c in 1:length(modNames)){
  module =modNames[c]
  moduleGenes2 = moduleColors==modNames[c]
  verboseScatterplot(abs(geneModuleMembership[moduleGenes2,c]), abs(geneTraitSignificance[moduleGenes2, 1]),
                      xlab = paste("Module Membership in", module, "module"), 
                      ylab = "Gene significance for psychosis",
                      main = paste("Module membership vs. gene significance\n"),
                      cex.main = 1.2, cex.lab = 1.2, cex.axis = 1.2, col = module,
                      pch=ifelse(abs(geneModuleMembership[moduleGenes2, c]) >= median(abs(geneModuleMembership)[moduleGenes2, c]) & abs(geneTraitSignificance)[moduleGenes2, ] 
                      >=min(abs(geneTraitSignificance)[moduleGenes2,]) , 19,1))
}; 
dev.off()


pdf("Supplementary_Figure_ModuleMembership_above_median_Psychosis.pdf",height=8,width=8)
for (c in 1:length(modNames)){
  module =modNames[c]
  moduleGenes2 = moduleColors==modNames[c]
  if(module %in% c("none")){
     
      verboseScatterplot(abs(geneModuleMembership[moduleGenes2,c]), abs(geneTraitSignificance[moduleGenes2, 1]),
                      xlab = paste("Module Membership in", module, "module"), 
                      ylab = "Gene significance for psychosis",
                      main = paste("Module membership vs. gene significance\n"),
                      cex.main = 1.2, cex.lab = 1.2, cex.axis = 1.2, col = module,
                      pch=ifelse(abs(geneModuleMembership[moduleGenes2, c]) >= median(abs(geneModuleMembership)[moduleGenes2, c]) & abs(geneTraitSignificance)[moduleGenes2, ] 
                      >=median(abs(geneTraitSignificance)[moduleGenes2,]) , 19,1))
  } else {
      verboseScatterplot(abs(geneModuleMembership[moduleGenes2,c]), abs(geneTraitSignificance[moduleGenes2, 1]),
                      xlab = paste("Module Membership in", module, "module"), 
                      ylab = "Gene significance for psychosis",
                      main = paste("Module membership vs. gene significance\n"),
                      cex.main = 1.2, cex.lab = 1.2, cex.axis = 1.2, col = module,
                      pch=ifelse(abs(geneModuleMembership[moduleGenes2, c]) >= median(abs(geneModuleMembership)[moduleGenes2, c]) & abs(geneTraitSignificance)[moduleGenes2, ] 
                      >=min(abs(geneTraitSignificance)[moduleGenes2,]), 19,1))
  }


}; 
dev.off()


```


```{r}
##annot = read.table("GAP_DL_FINAL.eset_final_gene_expression_matrix.txt",sep="\t", header=TRUE);
##dim(annot)
##annot$nuID
##str(annot)

annot<- fData(GX_lumi_object)
dim(annot)
names(annot)

probes = colnames(t_gene_expression_matrix)
probes2annot = match(probes, annot$TargetID)
# The following is the number or probes without annotation:
sum(is.na(probes2annot))
# Should return 0.

```

```{r}
# Create the starting data frame
geneInfo0 = data.frame(nuID = probes,
    geneSymbol = annot$SYMBOL[probes2annot],
    LocusLinkID = annot$ENTREZ_GENE_ID[probes2annot],
    moduleColor = moduleColors,
    geneTraitSignificance,
    GSPvalue)
# Order modules by their significance for psychosis
modOrder = order(-abs(cor(MEs, psychosis, use = "p")));
# Add module membership information in the chosen order
for (mod in 1:ncol(geneModuleMembership))
{
oldNames = names(geneInfo0)
geneInfo0 = data.frame(geneInfo0, geneModuleMembership[, modOrder[mod]],
MMPvalue[, modOrder[mod]]);
names(geneInfo0) = c(oldNames, paste("MM.", modNames[modOrder[mod]], sep=""),
paste("p.MM.", modNames[modOrder[mod]], sep=""))
}
# Order the genes in the geneInfo variable first by module color, then by geneTraitSignificance
geneOrder = order(geneInfo0$moduleColor, -abs(geneInfo0$GS.psychosis));
geneInfo = geneInfo0[geneOrder, ]


```

```{r}
setwd(topdir)
setwd(P2_output_dir)

write.csv(geneInfo, file = "Supplementary_table_6_WGCNA_geneInfo.csv")
setwd(topdir)
```
