==> R/p3_ENRICHMENT_05_11_16.Rmd <==
---
title: "WGCNA_GAP_Enrichment"
author: "DJL"
date: "07/07/2016"
output: html_document
---

#load libraries and set workind directory
```{r}
#libraries
rm(list = ls())
dev.off()
library(reshape)
library(WGCNA)
library(flashClust)
library(data.table)
library(dplyr)
<<<<<<< HEAD
=======
library(tidyr)
>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2

#directories
top_dir <-getwd()
data_dir <-"./data/"
P1_output_dir <-"./P1_Diff_Ex/output/"
P2_output_dir <-"./P2_WGCNA/output/"
P3_output_dir <-"./P3_Enrichment/output/"

```


#Functions
```{r Functions}

#function to remove uneccesary lists, by overlap with background. 
smaller_enrichmentlists<-function(enrichmentlist,Background,overlap){
    print("number of categories at start")
    print(length(unique(enrichmentlist$Eclass)))
    #add column indicating presence in background.
    print("col1")
    enrichmentlist$InGAP<-enrichmentlist[,1]%in%Background
    #Select only True rows
    print("True")
    dataTrue<-enrichmentlist%>%group_by(Eclass)%>%filter(InGAP==TRUE)
    #Table of Eclass with False and True numbers
    print("Eclass")
    dataTrue<-as.data.frame(table(dataTrue$Eclass))
    #find Eclass list with at least X (overlap) True probes.
    print("Eclass2")
    smalllist<-droplevels(dataTrue[dataTrue$Freq>=overlap,1])
    print("number of categories at end")
    print(length(smalllist))
    #Make reduced list
    return(enrichmentlist[enrichmentlist$Eclass%in%smalllist,1:2])
}

<<<<<<< HEAD
=======

##Function for adding column with list type. Only works with data input I create myself. Userlistenrichment internal lists probably wont work. 
New_Types_danlists<-function(Data){
  splitD<-strsplit(Data$UserDefinedCategories,"__")
  new_Category<-unlist(lapply(splitD, `[[`, 1))
  new_Type<-unlist(lapply(splitD, `[[`, 2))
  Data$UserDefinedCategories<-new_Category
  Data$Type<-new_Type
  return(Data)
}



>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2
```



#Enrichment data modification Adjust to data that is available. 
```{r}
#load background data
all_probes_file<-"Supplementary_Table_1_all_probes_5797_limma_results_AgeSexEthnicity_Adj.tsv"
all_probes<-read.csv(paste(P1_output_dir,all_probes_file,sep=""),sep="\t",header=TRUE)


#define lists
lists<-c("GO_Biological_Process_2015.csv","GO_Cellular_Component_2015.csv","GO_Molecular_Function_2015.csv","Kegg2016.csv")

#read data and perform functions and write to output dir.  
for(listname in lists){
  inputdata<-read.csv(paste(data_dir,listname,sep=""))
  listdata<-smaller_enrichmentlists(inputdata,all_probes$TargetID,20)
  write.csv(listdata,file=paste(P3_output_dir,"GAP_reduced_",listname,sep=""),row.names=F)  
}

```

## User List Enrichement for Diffferentially expressed genes
```{r gene_list}
#load background data
<<<<<<< HEAD
all_probes_file<-"Supplementary_Table_1_all_probes_5797_limma_results_AgeSexEthnicity_Adj.tsv"
all_probes<-read.csv(paste(P1_output_dir,all_probes_file,sep=""),sep="\t",header=TRUE)


#Subset to pvalue 0.05
all_probes$Groups <-all_probes$adj.P.Val <= 0.05
all_probes$Groups<-replace(all_probes$Groups, all_probes$Groups==FALSE, "background")
all_probes$Groups<-replace(all_probes$Groups, all_probes$Groups==TRUE, "FDR_Pass")
=======

all_probes_file<-"Datatoplot.tsv"
#all_probes_file<-"Supplementary_Table_1_all_probes_5797_limma_results_AgeSexEthnicity_Adj.tsv"
all_probes<-read.csv(paste(P1_output_dir,all_probes_file,sep=""),sep="\t",header=TRUE)


table(all_probes$Sig_LogFC_probes)
#Subset to pvalue 0.05
#all_probes$Groups <-all_probes$adj.P.Val <= 0.05
all_probes$Groups<-replace(all_probes$Groups, all_probes$Sig_LogFC_probes=="BACKGROUND", "background")
all_probes$Groups<-replace(all_probes$Groups, all_probes$Sig_LogFC_probes=="Diffexprs", "FDR_Pass")
>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2
setwd(top_dir)
setwd(P3_output_dir)
enrichments = userListEnrichment(all_probes$TargetID,all_probes$Groups,fnIn=c("GAP_reduced_GO_Biological_Process_2015.csv","GAP_reduced_GO_Cellular_Component_2015.csv","GAP_reduced_GO_Molecular_Function_2015.csv","GAP_reduced_Kegg2016.csv"),catNmIn=c("GO_BP","GO_CC","GO_MF","KEGG_2016"))
setwd(top_dir)

enpv<-enrichments$pValue
enrichment_probes<-unlist(lapply(enrichments$ovGenes,paste, collapse =";"))
enpv$Genes<-enrichment_probes
enpv<-enpv[order(enpv$InputCategories,enpv$Pvalues),]
enpvDT<-data.table(enpv)
enpvDT<-enpvDT[Pvalues < 0.05 & NumOverlap > 10,.SD[],by=InputCategories]

write.csv(as.data.frame(enpvDT),file=paste(P3_output_dir,"Supplementary_Table_9_User_list_enrichment_diff_expression_results.csv",sep=""),row.names=F)

```

##Userlist enrichment for WGCNA modules

##Subset to genes above median module membership.
```{r gene_list}
setwd(top_dir)
setwd(P2_output_dir)
## load WGCNA data
geneInfo<-read.csv("Supplementary_table_6_WGCNA_geneInfo.csv")
setwd(top_dir)
str(geneInfo)


#Modules
intModules<-names(table(geneInfo[,5]))
Psychosis_corr<-intModules[]
Non_Psychosis_corr<-intModules[]



output_data<-data.frame()
##  All probes with >= median membership for their module, excluding the above and excluding psychosis.
for (module in Non_Psychosis_corr)
  {
  # Select module probes
  modGenes = (geneInfo$moduleColor==module)
  # Get the module genes
  sgeneInfo<-(geneInfo[modGenes, ])
  #Select genes upregulated (above median in terms of module membership)
  modcol<-paste("MM.",module,sep="")
  genesf<-sgeneInfo[sgeneInfo[,modcol] >= median(sgeneInfo[,modcol]),c(1,5)]
  # Add all modules to dataframe sequentially
  output_data<-rbind(genesf,output_data)
  
}

table(output_data[,2])



background_probes<-geneInfo[!geneInfo[,1]%in%output_data[,1],c(1,5)]
background_probes[,2] <- "background"
Output_w_background<-rbind(background_probes,output_data)

str(Output_w_background)
table(Output_w_background[2])

setwd(top_dir)
setwd(P3_output_dir)
fileName = paste("Probes_median_membership_for_modules.csv", sep="");
write.table(as.data.frame(Output_w_background), file = fileName, row.names = FALSE, col.names = T,quote=FALSE,sep=",")


```




## User List Enrichement with KEGG and GO for WGCNA

```{r gene_list}


## Enrichment calculation. May take a couple of minutes. 

setwd(top_dir)
setwd(P3_output_dir)
enrichments_modules = userListEnrichment(Output_w_background[,1],Output_w_background[,2],fnIn=c("GAP_reduced_GO_Biological_Process_2015.csv","GAP_reduced_GO_Cellular_Component_2015.csv","GAP_reduced_GO_Molecular_Function_2015.csv","GAP_reduced_Kegg2016.csv"),catNmIn=c("GO_BP","GO_CC","GO_MF","KEGG_2016"))
setwd(top_dir)

##significant overlaps
enrichments_modules$sigOverlaps

enrichments_modules$sigOverlaps[order(enrichments_modules$sigOverlaps$InputCategories),]

##add genes to data.frame.
enpv_mm<-enrichments_modules$pValue
enrichment_probes_modules<-unlist(lapply(enrichments_modules$ovGenes,paste, collapse =";"))
enpv_mm$Genes<-enrichment_probes_modules

<<<<<<< HEAD
=======

>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2
##order by module and pvalue.
enpv_mm<-enpv_mm[order(enpv_mm$InputCategories,enpv_mm$Pvalues),]

##Make Datatable
enpv_mmDT<-data.table(enpv_mm)
##Select top 20, add pvalue < 0.05
enpv_mmDT<-enpv_mmDT[Pvalues < 0.05 & NumOverlap > 5,.SD[1:20],by=InputCategories]
#enpv_mmDT<-enpv_mmDT[,.SD[1:20],by=InputCategories]
<<<<<<< HEAD
str(enpv_mmDT)
=======

>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2

enpv_mmDF<-as.data.frame(enpv_mmDT)
enpv_mmDF<-enpv_mmDF[complete.cases(enpv_mmDF),]

table(enpv_mmDF$InputCategories)

getwd()
setwd(top_dir)
setwd(P3_output_dir)
write.csv(enpv_mmDF,file="Supplementary_Table_12_User_list_enrichment_WGCNA_results.csv",row.names=F)
setwd(top_dir)
<<<<<<< HEAD
=======


enpv_mmDF_type<-New_Types_danlists(enpv_mmDF)
getwd()
setwd(top_dir)
setwd(P3_output_dir)
write.csv(enpv_mmDF_type,file="Supplementary_Table_12_User_list_enrichment_WGCNA_results.csv",row.names=F)
setwd(top_dir)




##add genes to data.frame.
enpv_mm<-enrichments_modules$pValue
enrichment_probes_modules<-unlist(lapply(enrichments_modules$ovGenes,paste, collapse =";"))
enpv_mm$Genes<-enrichment_probes_modules


##order by module and pvalue.
enpv_mm<-enpv_mm[order(enpv_mm$InputCategories,enpv_mm$Pvalues),]

enpv_mmf<-enpv_mm%>%filter(CorrectedPvalues <= 0.05)
enpv_mmf_type<-New_Types_danlists(enpv_mmf)

#Make table for table
Paper_table_colnames<-c("Module","Library","Enriched Categories")
Table_paper <- data.frame(matrix(NA,ncol=3))
names(Table_paper)<-Paper_table_colnames
rowindex = 1
#group by color, > group by library >
for (module in unique(enpv_mmf_type$InputCategories)){
  modulesub<-filter(enpv_mmf_type,InputCategories==module)
  for(Type_cat in unique(modulesub$Type)){
    typesub<-filter(modulesub,Type==Type_cat)
    Usercategories_temp<-paste(typesub$UserDefinedCategories, sep="", collapse="; ")
    Category<-cbind(typesub[,c(1,3)],Usercategories_temp) 
    Table_paper[rowindex,] <-Category
    rowindex = rowindex + 1
    }
  
}

Table_paper[,1:2]

setwd(top_dir)
setwd(P3_output_dir)
write.csv(Table_paper,file="Table_4_User_list_enrichment_WGCNA_results.csv",row.names=F)
setwd(top_dir)


enpv_mmDF_type%>% group_by(InputCategories & Type)

enpv_mmDF_type$UserDefinedCategories
#collapse categories with seperator.

>>>>>>> a2d4c27737617d68932cf318180116b2fb1f39c2
```
